refactor: 단말(termanal)만 처리하도록 변경
