+++ emulator/src/main/java/com/pickcar/application/EmulatorService.java
@@ -2,7 +2,7 @@ package com.pickcar.application;
 
 import com.pickcar.domain.Terminal;
 import com.pickcar.infrastructure.EmulatorRepository;
-import com.pickcar.presentation.dto.EmulatorRequest;
+import com.pickcar.presentation.dto.request.EmulatorRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -12,19 +12,7 @@ public class EmulatorService {
 
     private final EmulatorRepository emulatorRepository;
 
-    public void on(EmulatorRequest request) {
-        Terminal terminal = Terminal.builder()
-                .mdn(request.getMdn())
-                .carId(request.getCarId())
-                .terminalId(request.getTerminalId())
-                .manufactureId(request.getManufactureId())
-                .packetVersion(request.getPacketVersion())
-                .deviceId(request.getDeviceId())
-                .build();
-        emulatorRepository.save(terminal);
-    }
-
-    public void off(EmulatorRequest request) {
+    public void terminal(EmulatorRequest request) {
         Terminal terminal = Terminal.builder()
                 .mdn(request.getMdn())
                 .carId(request.getCarId())
