+++ emulator/src/main/java/com/pickcar/presentation/controller/EmulatorApiController.java
@@ -1,12 +1,9 @@
 package com.pickcar.presentation.controller;
 
 import com.pickcar.application.EmulatorService;
-import com.pickcar.presentation.dto.EmulatorRequest;
-import java.util.Map;
-import java.util.Objects;
+import com.pickcar.presentation.dto.request.EmulatorRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -18,18 +15,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class EmulatorApiController {
 
-    private final EmulatorService emulatorService;
+    EmulatorService emulatorService;
 
-    @PostMapping("/engine/on")
-    public void emulatorEngineOn(@RequestBody EmulatorRequest request) {
-        log.info("POST /api/v1/engine/on - EmulatorRequest: {}", request);
-        emulatorService.on(request);
+    @PostMapping("/terminal")
+    public void emulator(@RequestBody EmulatorRequest request) {
+        log.info("POST /api/v1/engine - EmulatorRequest: {}", request);
+        emulatorService.terminal(request);
     }
 
-    @PostMapping("/engine/off")
-    public void emulatorEngineOff(@RequestBody EmulatorRequest request) {
-        log.info("POST /api/v1/engine/off - EmulatorRequest: {}", request);
-        emulatorService.off(request);
-
-    }
 }
